CLEANUP: ceba: remove the generic dup-aware versions
By definition this model cannot support duplicates since it indexes the
node's address. This is also reflected by the fact that both the unique
and generic versions are strictly identical and that there was no
next_unique() nor next_dup() functions. Let's just get rid of the generic
functions and stick to the unique ones. At least their name reflect this
specificity.