BUILD: only include headers during development
The risk to keep some possible API compatibility issues was addressed by
commit 29a9665 ("BUILD/MINOR: fix missing includes that hide incorrect
APIs"), but it causes one problem which is that these files must always
be located in the same directory as the C file, a constraint which is
not desirable for integration into other projects. We don't need these
include files anyway in release mode, so let's just include them when
CEBTREE_DEV is defined. This is sufficient to spot issues during
development and doesn't impose any integration constraint.