DOC: naming: rename the "large" relative addressing model to "huge"
The reason is the use of the "l" which conflicts with the "long" key
type. "huge" is still quite explicit and doesn't conflict.

With cebtree, relative addressing is relatively easy, it can always
be relative to the root and multiplied by the node size, because the
root will never point to itself (i.e. NULL remains valid), and all
operations require a reference to the root. This will mean that:
  - 16-bit pointers (small) will support addressing up to a 256kB area
  - 32-bit pointers (medium) will address up to 32GB
  - 64-bit pointers (huge) will address up to 256PB

Also, if indirect nodes are needed (is, im etc), then the total node
size will be 3 pointers (left,right,key). Either a hole will be left,
or the multiplication can be done with that value, or alternately by
a single pointer size, hence cut the sizes above in half.