MEDIUM: integers: make all integer functions include _ceb_int.c instead
There used to be a few glitches due to copy-pastes recently (e.g.
64-bit key placed into the 32-bit arg etc). Let's reduce code duplication
and make the 32/64/ulong versions just rely on _ceb_int.c instead. This
way there's a single file to change in the future for all 3 types. It was
verified that the resulting object files are exactly the same.