MINOR: block: centralize direct and indirect blocks into _ceb_blk.c
For the same reasons as for integers, we're just dealing with exact
code duplication between direct and indirect blocks. Let's just move
that to _ceb_blk.c. Note that this cannot be merged with strings
because the API differs: for blocks, functions take the block length
while for strings it's supposed to be zero-terminated.