MEDIUM: tree-wide: allow to disable the dump code
The dump code represents ~1/3 of the total size and is only used during
development or specific debugging. In addition it comes with dependencies
on stdio. All this stuff is not necessarily desired for the application
code, so let's not build it by default. It's now conditioned to
CEB_ENABLE_DUMP, which is added by default by the makefile. This way,
when the code is included into another project, it will not be built
by default. It's also possible to drop support for it when building a
library using the current makefile by just setting "WITH_DUMP=".