DOC: naming: suggest how to name the variants that include a key offset
One of the benefits of the current storage system is that the node doesn't
break structure alignment, and that it can make sense to place a key
anywhere else, including before the node or much further to reuse an
existing one. The functions dealing with this currently have "_ofs"
in their name, but this could be generalized with the "o" placed
between the addressing mode and the key. It can also help name files.