CLEANUP: tools: drop the tools.h file entirely
Most of what was in it was only used by ebtree but we don't need it
here anymore since the few relevant parts were adopted by cebtree-prv.h.
Let's just get rid of the file to simplify the adoption of the lib.
The makefile was updated to remove the search path in the common dir.