OPTIM: private: support comparing one word at a time in string_equal_bits()
Perf profiles clearly show 70% of the time spent in that function. Let's
implement the ability to compare one word at a time on architectures that
support unaligned accesses. It's a bit tricky because we don't want to
cross a page boundary since we don't know where the trailing zero is
located. Most of the time, though, strings are small and don't start in
the last bytes of a page. For this reason, the solution taken here is to
measure how far each string is from the next page boundary and to read
at most that many words. After that, we fall back to one byte at a time.
We need to disable this when -fsanitize=address is in use of course.

This results in a 8% speedup of the various speed tests involving strings,
but the code got quite bigger.