BUG/MEDIUM: fix the prev() implementation with duplicates
The condition on prev() was incorrect due to a line-by-line application
of the tests in the comments. Actually both situations that permit to
return from->b[0] require the element to be a dup. Otherwise we must
restart from the restart pointer. Due to this, the reverse scan would
loop forever if a tree supporting duplicates was used without dups.

This was reproducible using ./tests/testceb32 3 2 1