BUG/MAJOR: private: fix deletion of the first dup entry
The node parent's down pointer was adapted to point to the new last
node in the case we'd delete the first duplicate of a series (case 1 in
the comments), which is a leaf whose node part necessarily is the parent
on top of the dup list, but that's not actually correct, as we're
replacing the node on top of us which is already the parent, so it's
actually the gparent that has to be fixed in this case. This is also
reflected in the second part where it's always done.

This was easily reproduced with ./tests/testceb32 3 3 2 2 1 1 which
would segfault during the ceb32_first() that followed the first delete
due to the NULL left in the removed node's p[0] by delete.