WIP: now support SSE2-based comparisons
SSE2 is mandatory on any x86_64, so it looks like a good plan to support
it. However it turns out that in the end it's slower than the 64-bit int
version (3.12s). Maybe it's the mix of int and sse ops that's not optimal.
Some areas could definitely be improved though, for example we continue to
rely on bswap while sse probably has better. Also x86_64 seems to always
support MOVBE that would allow to read directly from the RAM in big
endian format. Those are still points to consider. It seems however
that the xbit vs zbit checks could be reused for the other mode.