WIP: slightly faster zero search (2.78)
Now the zero search is slightly faster, and it doesn't require to check
the bit of r yet, which allows us to only check this bit when we're
certain r is non-null, thus removing a check on the most common return
path.

Compared to the non-vector 3.617, it's 30% faster for a significantly
larger footprint (80kB vs 47kB for a test program).