MAJOR: api: type node pointers to prepare the transition to tagged pointers
Now we have roots which are tagged pointers, and regular nodes. The
function __ceb_dotag() turns a node into a tagged pointer, __ceb_untag()
does the opposite, and __ceb_clrtag() clears the tag regardless of its
value. For now actually __ceb_clrtag() just returns the original value.

Tags were always set to zero for now, so when starting to use them, one
must be extremely careful about the expected values. It's also worth
mentioning that the dump functions tend to turn roots to nodes a bit
quickly and that maybe a copy of the root's tag will still be needed
during the descent.

Callers will now have to use ceb_root instead of ceb_node* for their
root.

At the moment the code continues to work.