MEDIUM: private: tag pointers to the leaves
From now on, a pointer to a leaf will have its lowest bit set. This will
permit to distinguish them without having to re-read the whole key. One
of the expected benefits is to be able to restart comparing strings from
the last verified offset instead of re-reading them all from the beginning
to detect variations between the split point.

This is also valid for duplicate sub-trees, which are still considered as
pure nodes, so that all their pointers are untagged except the leftmost
one which points to the leaf.

This ensures very little pointer manipulation overall.