OPTIM: private: slightly unroll the string matching loop
After numerous attempts (including vectorizing), it turns out that the
most effective opitimization remains to just unroll the loop twice. It
does not add too much code and doesn't come with an expensive setup for
each level. This improves the overall performance by 3.2%, making the
testcebus test down to 3.14s for 20k loops and 15.67s for 100k loops,
which means 80M lookups in 12.53s or 6.385M lookups/s.