MEDIUM: descent: make use of the knowledge of leaves to simplify int matching
Since we know that the currently visited node is a leaf, we can interrupt
the descent earlier. The real benefit here is no longer having to set/check
pxor32/64, thus one less variable in the hot path. The variable is still set
because it's used in the dbg() statements. It's simply marked unused so that
the compiler does not complain about it being set while not being used. The
performance on ulong jumped from 19.0 to 20.5 M lookups per second, or 7.7%.