CLEANUP: descent: merge the direct and indirect blocks/strings
The code for direct and indirect branches only differs by the use of
l->ptr vs l->str or l->mb depending on the type. All the rest is strictly
the same, so let's reduce code duplication and merge these branches by
using only a ternary expression to retrieve the pointer (which would be
resolved by the compiler to the same pointer anyway in the end). This
allows to remove 135 lines.