CLEANUP: descent: do not needlessly compare matching lengths
When going down for a delete(), we compare the node in order to figure
where we'll have to mark a node's parent and stop. This requires some
extra comparisons. However, these are done in a way that is more complicated
than needed, because the tree structure offers the guarantee that if there
are at least X bits in common between the two sides, then all of these bits
are also present in the current node's key. And since the delete() operation
applies to nodes in the tree, we're certain the key is there. As such, the
test on mlen, rlen, llen and xlen is pointless, because mlen will always be
set to xlen. Let's just simplify this and perform the comparison with xlen
bits, which also makes it more logical to check for it to be infinite.

This extra test was added when planning to add the pick() operation, where
the key does not necessarily belong to the tree, but since we're now only
deleting entries present in the tree the problem does not exist.