OPTIM: private: improve flsnz() calculation in string_equal_bits()
The following expression on the critical path is difficult for the
compiler:

    (beg << 3) - flsnz(c)

First, the compiler doesn't know that flsnz() is comprised between
1 and 8. Second, subtracting is always a difficult operation on x86
as it requires a register swap, which is sad considering that the left
shift could be combined with an addition.

Let's just write it this way instead:

    (beg << 3) + ((flsnz(c) - 1) ^ 7) - 8;

While apparently more complicated, the -1 cancels the +1 that comes
from flsnz32(), so we're dropping two operations; on machines using
a clz operation, the ^7 is merged with ^31 which gives ^24, while on
those using bsr, we just keep ^7; the -8 becomes a static offset in
an LEA operation. So instead we're just doing:

    bsr eax, ecx                  # ecx=MSB(c)
    xor ecx, 7                    #
    lea eax, [beg * 8 + ecx - 8]  # eax = expression

The size is reduced by ~32 bytes per function, and the speedcebus test
comes from 3.070 to 3.000s, or 2.3% faster.

(cherry picked from commit 89b4ff180a99ce0efd5255304916af73811b0bcb)
Signed-off-by: Willy Tarreau <w@1wt.eu>