BUG/MAJOR: private: fix deletion of leaf in a dup list
We were incorrectly updating the grandparent's leaf instead of the
node's parent's leaf. In many tests they were the same, but in some
tests it was obvious there was a problem, for example in this one,
when deleting the first 0011:

   ./tests/testcebs -D 4 1 1 1000 0011 0011 0101

One of the reasons is that nparent was incorrectly updated to the
last list entry when going down a tree with a list element, due to
the comparison that matched entirely. Instead we must remember that
we have not yet found the node's parent and perform the comparison
only in this case.

(cherry picked from commit 880857acb1154120ebcecf34880f8b0914ebde4c)
Signed-off-by: Willy Tarreau <w@1wt.eu>