MINOR: descent: don't carry the nparent & npside values all along
We just need to store nparent and npside the very first time they are
met, and there's already a test to figure if we've met them. But both
ret_nparent and ret_npside point to a caller's variable so there's no
point in keeping another local variable all along the descent, let's
instead simply check the caller and assign it immediately when the
desired values are found. This way we can reduce the code by about 1.2%.