BUG/MAJOR: private: deleting dups leaves a few more corner cases
First, when deleting the leaf at the extremity of a dup list, the
list before it must be tagged to indicate a leaf. This was not done,
resulting in corrupted pointers. Second, the leaf parent, which is
the node on top of the list, may be the removed element. Normally
this is not a problem since the first of the list will replace it,
except that if the first was also the last, it will leave itself as
a leaf, and in this case it must also be tagged on its branch.

All of this seems to indicate that deletion with tagged nodes is more
complicated than it needs, probably because we don't tag list elements.
It looks like there would be some value in tagging lists as leaves.
This would make it easier to detect the leaf parent (the one with a tag
on the followed branch), and all permutations inside the list would be
easy. Also, the tag on the leaf in the dup list currently doesn't bring
any value since it's always at last->b[1]->b[0] anyway.