BUG/MAJOR: descent/string: do not compare negative string offsets
Strings report ~0 or -1 as the matching length to indicate strings that
do exactly match. However we only check for this when ret_is_dup is set,
as a proxy to detect the support for duplicates by the caller. But this
is inaccurate as some callers don't use it (e.g. ceb_last) and it ends
up continuing the descent with -1 as the offset to continue matching
strings, resulting in crashes or infinite loops. Other types have this
issue with no real impact, since they're not using the previous match
length as an offset to restart from. One place that fails to run the
proper check is the strcmp() performed on delete. If we've found the
right node after visiting its matching parent (e.g. at the top of a
dup list), the strcmp() will get a negative xlen offset and crash.
Let's properly check that xlen is positive there, indicating a full
match, hence that it's not necessary to check again.

The following tests segfaults before this fix:
     ./tests/testcebs -m 255 1000 1000 1

Note that a solution will still be needed for the ret_is_dup situation.

(cherry picked from commit cd75527e75279d1a2573c0f2b352a049ef3f140d)
Signed-off-by: Willy Tarreau <w@1wt.eu>