CLEANUP: descent: get rid of the "found" variable
It was meant as an optimization to avoid comparing the whole string a
last time to figure what branch we're going through, but it ends up
adding non-negligible cost at each stage to save very few at the end.
For example, this test that adds and looks up 1M pattern made of at
most 512k distinct keys:

   $ tests/testcebs -m 524287 1000000 1000000 1

takes 2.72s with the original code and 2.68 once cleaned up. In addition
the code is easier to read with this and doesn't convery obscure exceptions.

(cherry picked from commit 5f5879131e27e85782587ace53c2055aa8952f41)
Signed-off-by: Willy Tarreau <w@1wt.eu>