BUG/MAJOR: private: add an is_dup_ptr to almost all calls
Strings report ~0 or -1 as the matching length to indicate strings that
do exactly match. However we only check for this when ret_is_dup is set,
as a proxy to detect the support for duplicates by the caller. But this
is inaccurate as some callers don't use it (e.g. ceb_last) and it ends
up continuing the descent with -1 as the offset to continue matching
strings, resulting in crashes or infinite loops. Other types have this
issue with no real impact, since they're not using the previous match
length as an offset to restart from. In any case we should normally
simply stop descending here since it indicates a duplicate.

The following tests loop forever on this issue:
     ./tests/testcebl 1 1 1

Given that most callers already rely on is_dup_ptr, let's just continue
this way and add it to all functions which didn't have it (last,
lookup_lt/le, next/prev_unique etc). The only remaining exceptions are
ceb_next_dup() and ceb_prev_dup() which already handle it locally. This
is cheap because we don't even need to initialize the variable, it will
be set appropriately if needed.

(cherry picked from commit b284ad8a80f012de676c81d331760682e0c737b1)
Signed-off-by: Willy Tarreau <w@1wt.eu>