CLEANUP: descent: drop the forced cast to unsigned on the lengths
This was inherited from the early code when attempts were made with
signed and unsigned values. No need to cast llen, rlen nor xlen to
unsigned anymore, they're all size_t, and the cast has a tiny cost
for certain operations (depending on compiler optimizations). The
code is now 0.25% smaller.