MEDIUM: private: change the internal representation of lists to be tagged
By tagging lists, we're having something much clearer: is_leaf always
matches either leaves or lists, thus the leaf parent is unambiguously the
node above either a leaf or a list. We don't need to recheck ret_is_dup
on each step during the descent since a dup will always match a leaf.
The code is 4% smaller and should be easier to simplify or debug.