MINOR: descent: simplify the leaf test
leaves can be handled earlier now that we want to stop on the first of
a list or a leaf. We can stop at the very beginning of the loop as soon
as is_leaf matches, where previously we were only dealing with the
nodeless leaf.

The code, in addition to being a bit simpler, becomes 7.2% smaller and
slightly faster.