OPTIM: descent: detect the node's parent before entering the loop
We don't really need to re-compare each node's key with the currently
looked up one, in practice we can also do this during the descent when
comparing the current key with the branches since we have the correct
information there. However, this requires a preliminary check in case
the root itself is the node's parent because the looked up node is the
topmost one.

Tests have shown that complex types such as strings and blocks do benefit
from doing it once at the top and not repeating it inside the descent
loop, while simpler types strongly prefer to do it during the descent
since it's only a comparison.

This patch implements the nparent check at the root for blk and str,
and gets rid of it during the descent. The operation on such types is
now very slightly faster for delete, but more importantly the code is
significantly smaller (3-4kB saved for strings, a few tens of bytes for
blocks).