MINOR: descent: just compare node's leaf with a tagged root at the end
There's a mandatory test at the end of the loop on the new node to see
if it matches a current node's leaf. But we can simplify it since the
purpose is that it matches a leaf. By doing so we drop ~1.2kB per
implementation due to the simplification which possibly discourages the
compiler from over-engineering the loop.