OPTIM: descent: help the compiler knows the previous len is non-negative
For strings, the compiler cannot always know what conditions will be met
before calling string_equal_bits(). We can actually tell it that plen is
never negative, which helps it reduce the code by 12% and improve the
performance by 5%. Now on a skylake 4.4 GHz it takes 633ms to perform 1
million lookups among 1M strings, and 1.03s to do the same on a cortex
A720 at 3.0 GHz.