CLEANUP: private: move the dump functions to cebtree-dbg.c
The dump functions are large and static due to being located in
cebtree-prv.h. This doesn't make sense given that they have to support
all types and are replicated for each and every implementation.

Let's move this code to cebtree-dbg.c instead and leave their prototype
in cebtree-prv.h to ease usage. Now we only need to link with this file
to benefit from them, otherwise the code is simply not compiled. With
this change, each implementation was reduced by about 5kB, and even 6kB
if building without CEB_ENABLE_DUMP.