MINOR: descent: keep a copy of the is_leaf for each side
This avoids propagating the root pointers all along the function
and performing the check at the end. Instead the check is performed
early when the pointer is known and its result is used at the end.
The code is now ~20% smaller because the compiler manages to avoid
a number of variables assignment and merge half of them. The
performance about the same (~0.5% slower).