MINOR: private: improve string_equal_bits to optimize the loop
By moving the loads earlier and pre-loading the next cycle, we
can speed the function up a little bit (~1%) and further reduce
the code by about 20%.

Attempts were made at loading 3 and 4 bytes per iterations, but
there are diminishing returns and it only makes sense when the
function is not inlined. However, there's a 1-2% performance
loss when it's not inlined so it's not worth the pain. But in
this case the code is further reduced, which could be a possible
choice based on the code optimization.