BUG/MAJOR: private/delete: properly attach first after replacement
When deleteing the leaf in a 3+ dup list, the first list element needs
to be moved to replace the node and leaf parts. Its branches must then
point to the *last* pointer, like the previous node did, not the first,
otherwise we instantly lose other elements. And it turns out that the
previous assignment of first branches already did the job by picking
the removed node's elements. We know that it's OK because we're deleting
a leaf from a list (checked by first->b[0]), thus by definition the node
did not point to itself and the leaf wasn't pointed to by anything but
the first list element which is going to replace it, meaning that by
definition we're not leaving any dangling pointers.