OPTIM: descent: move the prefetch earlier
Moving the prefetch at the end of the previous loop improves the
performance by 13% for ins/del and 21% for pure lookups on x86. And
keeping one first lookup for deletions before the loop gets one
extra percent. This is not much surprising because the prefetch was
already done very close to the place where the values were going to
be used.