OPTIM: descent: stop at the node when the path to the leaf is not used
For plain lookups, we don't necessarily need to go down to the leaf,
and we can equally stop at the node when it matches. It turns out
that most of the time we already have this info (often from the
previous node) since the key has already been compared to each side's
keys. So now when we know that the caller doesn't request any special
info, we just stop at the node.

This brings an overall 20% performance gain for read-mostly workloads,
and string lookups even become slightly faster than in ebtree.