TESTS: bench: support choosing the series of keys to be used
When using timers, keys are mostly grouped. When using hashes, they're
normally very well distributed. And finally there is the special case
of the worst case, to check what happens for example in case of keys
generated for the purpose of attacking the system by producing long
chains. We now have a new argument -S which allows to choose between
all of these. The default is zero and is the pure random. Note that
previously, the default setting was to have longest length for leading
zeroes and randoms for leading ones, which did not represent anything
realistic. Here the differences between the 3 modes are more visible.

(cherry picked from commit 6c7fe34e81f3d2e26293c3c66f63d2b924375b38)
Signed-off-by: Willy Tarreau <w@1wt.eu>