OPTIM: descent: rearrange the prefetching at more optimal places
The prefetching was performed either for reading or writing depending
on the operation, and at a place that was not optimal. It turns out
that moving it earlier in the loop for reads only, and adding a few
write-prefetches at the end when returning pointers that are likely
to be modified yields ~3% better performance both for pure writes and
pure lookups while simplifying the logic in the loop, so let's do this.

(cherry picked from commit 0ee1e615bbc64a8ac4bb777736b3f2c329bd0b16)
Signed-off-by: Willy Tarreau <w@1wt.eu>