MINOR: api: add a _key() operation for each type to return a pointer to the key
Sometimes it would be convenient to directly get a pointer to the key
itself, especially in loops covering ranges. This commit adds ceb*_key()
that takes a node and returns either NULL or a pointer to a valid key.
It also takes care of dereferencing indirect keys (cebib, cebis) so as
to ease memcmp() / strcmp(). The doc was updated with an example.