BUG/MEDIUM: private: never read past end in string_equal_bits()
The change in commit 09902b8 ("MINOR: private: improve string_equal_bits
to optimize the loop") accidentally resulted in reading byte N+1 before
knowing if byte N was zero or not, which was caught by ASAN.

Let's partially undo that fix while still checking early if we can leave
the loop. In the end the code is now even 4% faster on older compilers
(gcc-9.5 and below) while also quite a bit larger. On newer compilers
(11 and above) which always produce significantly slower code than old
ones, there's no observable difference, so the code was written in a
way that the compiler will have no problem optimizing, giving a good
balance of performance on all of them.

It's interesting to note that by just assigning "c" first before
entering the loop then at the end of the loop, the code becomes
much bigger on older compilers and significantly faster at the same
time (~1%), but this approach doesn't work on new ones.