MINOR: descent: let the compiler know that there are no NULLs in the tree
Adding a __builtin_unreachable() test on the next node allows the compiler
to know that there are no NULLs in the tree, and simplify some of the
upper functions which test for NULLs, so that their test directly matches
the explicit null returns. The test programs are reduced by around 1%,
and most functions got smaller by approx 16 bytes, indicating one or
two less tests. The multi-thread cebul test shows a 0.9% perf gain.