MINOR: api: provide macros to work directly from items and struct members
The new cebXX_item_YY() macros map the cebXX_ofs_YY() functions to work
based on the container item instead of the node. This way the user does
not need to know about ceb_node, and it becomes trivial to use split
node members and key members. It is also convenient to navigate between
items using the prev/next etc.

The type of the item is implicitly detected from the item pointer itself,
except for the few operations which do not take one (first/last/lookup/pick),
in which case it's always passed last as a reminder that it's the type of
the returned pointer.

The doc was updated to reflect this addition.