MINOR: api: mark as const the roots for lookup functions
All functions except insert/delete/pick do not change the contents of
the tree. It's pointless to mark the root pointer as mutable there, as
it may become unconvenient to deal with in certain scenarios where the
root pointer would be passed from a const struct.

Let's instead declare the root as "struct ceb_root *const *root" for all
these functions. There's no risk of misuse anyway since all returned
nodes will need the root to perform modifications and using modifying
functions on such roots will then be detected.

This change remains backwards compatible with existing code, and it was
verified that the resulting code is exactly the same.