API: make the immediate mode explicit via _imm
It's really problematic when migrating code from ebtree to cebtree that
the default mode uses an implicit key offset, because a mistake is quickly
done and hard to spot. Better ditch the "default" API and make it take an
explicit "_imm" prefix to indicate that the developer actually *wants* to
use the immediate API.

This *will* break code working with previous versions. "_imm" will have to
be inserted where it was not. The following "sed" rule was used to migrate
everything at once and should perform the change well in existing code:

    s,\b\(ceb[^_]\+_\)\(ins\|nex\|del\|def\|key\|[flp]\),\1imm_\2,g