MINOR: api: add ceb_init_root() and ceb_init_node()
Initializing a root is not particularly complicated, but initializing
a node can be less intuitive when it's part of an existing structure,
and one doesn't necessarily want to memset() the whole node. Thus let's
add two functions to perform such initializations for the sake of keeping
the calling code clean and readable.